I don't know what to do

+++ src/cn/hnu/eg/sys/Message.java
@@ -1,7 +1,7 @@
 package cn.hnu.eg.sys;
 
 public interface Message {			
-	public int toValue();	
+	public double toValue();	
 	public boolean isHalt();
 	public boolean isActive();
 	public boolean isDeath();
